feat: Implement KYC controller with basic details submission, risk profiling, capital management, experience submission, and status retrieval
feat: Add signal controller with CRUD operations for signals, bulk creation, and user-specific signal retrieval

feat: Create subscription controller for managing bot subscriptions, including creation, retrieval, cancellation, and deletion
